restaurant_os/requirements/base.txt
appdirs==1.4.4
argcomplete==3.6.3
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==20.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.13
importlib_resources==7.1.0
numpy==2.4.4
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.12.1
python-barcode==0.16.1
python-decouple==3.8
python-escpos==3.1
PyYAML==6.0.3
qrcode==8.2
requests==2.33.1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
